Move all DynDynTable-related items into a new table module
Previously, many of the pieces that ultimately were part of representing
DynDynTable were in the base module or scattered about. For better
organization, a new table module has been created for them.
