+++ dyn-dyn/src/lib.rs
@@ -17,6 +17,13 @@
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
+mod dyn_trait;
+mod fat;
+mod table;
+
+#[doc(hidden)]
+pub mod internal;
+
 /// Declares a trait as being a base trait for downcasting.
 ///
 /// This macro marks a trait as being a base for dynamic trait object downcasting. All `impl` blocks for this trait will need to use the
@@ -115,134 +122,19 @@ pub use dyn_dyn_macros::dyn_dyn_cast;
 pub use dyn_dyn_macros::dyn_dyn_derived;
 
 pub use fat::DynDynFat;
-
-mod dyn_trait;
-mod fat;
-
-#[doc(hidden)]
-pub mod internal;
+pub use table::{DynDynTable, DynDynTableEntry, DynDynTableIterator};
 
 #[cfg(doc)]
 use core::ops::Deref;
 
 use cfg_if::cfg_if;
-use core::any::TypeId;
-use core::fmt::{self, Debug};
 use core::marker::{PhantomData, Unsize};
 use core::ops::DerefMut;
 use core::ptr::{DynMetadata, NonNull};
 use stable_deref_trait::StableDeref;
 
-use crate::dyn_trait::{AnyDynMetadata, DynInfo, DynTrait};
+use crate::dyn_trait::DynTrait;
 use internal::*;
-
-/// An entry in a concrete type's table of downcast-exposed traits.
-///
-/// Each entry represents a single trait object that the concrete type in question can be downcast to. Note that entries will only appear
-/// for bare trait object types, i.e. `dyn Trait`. Trait objects with extra marker types, e.g. `dyn Trait + Send`, are handled specially
-/// by the [`dyn_dyn_cast!`] macro and do not appear in a concrete type's trait table.
-pub struct DynDynTableEntry {
-    ty: DynInfo,
-    meta: AnyDynMetadata,
-}
-
-impl DynDynTableEntry {
-    #[doc(hidden)]
-    pub const unsafe fn new<
-        T,
-        D: ?Sized + ~const DynTrait + 'static,
-        F: ~const FnOnce(*const T) -> *const D,
-    >(
-        f: F,
-    ) -> DynDynTableEntry {
-        DynDynTableEntry {
-            ty: DynInfo::of::<D>(),
-            meta: D::meta_for_ty(f),
-        }
-    }
-
-    /// Gets the [`TypeId`] of the trait object corresponding to this entry.
-    pub fn type_id(&self) -> TypeId {
-        self.ty.type_id()
-    }
-
-    /// Gets a human-readable name representing the trait object corresponding to this entry.
-    #[cfg(feature = "dynamic-names")]
-    pub fn type_name(&self) -> &'static str {
-        self.ty.name()
-    }
-}
-
-impl Debug for DynDynTableEntry {
-    #[cfg(feature = "dynamic-names")]
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "DynDynTableEntry(<{}>: {:?})",
-            self.type_name(),
-            self.meta
-        )
-    }
-
-    #[cfg(not(feature = "dynamic-names"))]
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "DynDynTableEntry({:?}: {:?})", self.type_id(), self.meta)
-    }
-}
-
-/// A table of trait object types that a concrete type can be downcast to.
-#[derive(Debug, Clone, Copy)]
-pub struct DynDynTable {
-    traits: &'static [DynDynTableEntry],
-}
-
-impl DynDynTable {
-    /// Finds the metadata corresponding to the type with the provided [`TypeId`] in this table or `None` if no such metadata is present.
-    pub fn find_untyped(&self, type_id: TypeId) -> Option<AnyDynMetadata> {
-        self.traits
-            .iter()
-            .find(|&entry| entry.ty.type_id() == type_id)
-            .map(|entry| entry.meta)
-    }
-
-    /// Finds the metadata corresponding to the trait `D` in this table or `None` if no such metadata is present.
-    pub fn find<D: ?Sized + DynTrait + 'static>(&self) -> Option<DynMetadata<D>> {
-        self.find_untyped(TypeId::of::<D>()).map(|meta| {
-            // SAFETY: This metadata corresponds to the trait D, so we can downcast it
-            unsafe { meta.downcast() }
-        })
-    }
-
-    /// Returns a reference to the slice of entries in this table
-    pub fn into_slice(self) -> &'static [DynDynTableEntry] {
-        self.traits
-    }
-
-    #[doc(hidden)]
-    pub const fn new(traits: &'static [DynDynTableEntry]) -> DynDynTable {
-        DynDynTable { traits }
-    }
-}
-
-impl IntoIterator for DynDynTable {
-    type Item = &'static DynDynTableEntry;
-    type IntoIter = DynDynTableIterator;
-
-    fn into_iter(self) -> Self::IntoIter {
-        DynDynTableIterator(self.traits.iter())
-    }
-}
-
-/// An iterator returning all entries in a [`DynDynTable`].
-pub struct DynDynTableIterator(core::slice::Iter<'static, DynDynTableEntry>);
-
-impl Iterator for DynDynTableIterator {
-    type Item = &'static DynDynTableEntry;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-}
 
 /// Wraps a reference to a pointer implementing [`GetDynDynTable<B>`] and which can be dereferenced to perform the downcast.
 ///
